src/requirements.txt
python-frontmatter>=1.1.0
Jinja2>=3.1.2
Markdown>=3.4.3
PyYAML>=6.0.1
requests>=2.31.0
google-genai>=1.0.0
python-dotenv>=1.0.0
